requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.51.0
pdfplumber==0.11.4
python-docx==1.1.2
aiofiles==23.2.1
